[GEOT-6499] MBTilesDataStoreFactory declares it can process any JDBC param set, regardless of the dbType
